optimize: 🚀 优化数据结构

+++ walmart-sdk/src/main/java/io/github/ealenxie/walmart/marketplace/CurrencyAmountUnit.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 /**
  * @author create by sch  2023/7/11 10:15
  * @version 1.0
@@ -15,7 +17,7 @@ public class CurrencyAmountUnit {
      * currencyAmount
      */
     @JsonProperty("currencyAmount")
-    private Double currencyAmount;
+    private BigDecimal currencyAmount;
     /**
      * currencyUnit
      */
